refactor: split OverviewTab.tsx (946 lines) into 5 focused modules
Split into:
- OverviewTabTypes.ts (64) — types + constants
- OverviewTabHelpers.ts (330) — pure functions
- OverviewTabComponents.tsx (210) — small UI atoms
- OverviewSummaryTable.tsx (320) — main grid renderer
- OverviewTab.tsx (85) — shell component

18 existing tests pass. Codex: 4/4 PASS.
