+++ cloud/apps/web/src/components/analysis/tabs/OverviewTab.tsx
@@ -4,29 +4,17 @@
  * Displays a semantics-backed summary table above condition-level drilldowns.
  */
 
-import { useMemo, useRef, type ReactNode } from 'react';
-import { Info } from 'lucide-react';
-import { useNavigate } from 'react-router-dom';
 import type { PerModelStats } from './types';
 import type { AnalysisResult, VarianceAnalysis, VisualizationData } from '../../../api/operations/analysis';
 import type { Run } from '../../../api/operations/runs';
-import { Button } from '../../ui/Button';
-import { CopyVisualButton } from '../../ui/CopyVisualButton';
-import { Tooltip } from '../../ui/Tooltip';
-import { PairedRunComparisonCard } from '../PairedRunComparisonCard';
-import {
-  resolveScenarioAttributes,
-} from '../../../utils/decisionLabels';
 import {
   ANALYSIS_BASE_PATH,
   type AnalysisBasePath,
-  buildAnalysisTranscriptsPath,
 } from '../../../utils/analysisRouting';
 import type {
   AnalysisSemanticsView,
-  PreferenceViewModel,
-  ReliabilityViewModel,
 } from '../../analysis-v2/analysisSemantics';
+import { OverviewSummaryTable } from './OverviewSummaryTable';
 
 type OverviewTabProps = {
   runId: string;
@@ -49,855 +37,6 @@ type OverviewTabProps = {
   currentAnalysis?: AnalysisResult | null;
   companionRun?: Run | null;
 };
-
-type ConditionRow = {
-  id: string;
-  attributeALevel: string;
-  attributeBLevel: string;
-  scenarioIds: string[];
-};
-
-type ConditionRepeatStats = {
-  directionalAgreement: number | null;
-  medianSignedDistance: number | null;
-  neutralShare: number | null;
-  totalCount: number;
-  maxRange: number | null;
-};
-
-type RepeatPatternSource = {
-  runId: string;
-  varianceAnalysis: VarianceAnalysis | null | undefined;
-  conditionRows: ConditionRow[];
-};
-
-type RepeatPattern = 'stable' | 'softLean' | 'torn' | 'noisy';
-const REPEAT_PATTERN_LABELS: Record<RepeatPattern, string> = {
-  stable: 'Stable',
-  softLean: 'Soft Lean',
-  torn: 'Torn',
-  noisy: 'Unstable',
-};
-
-const SUMMARY_COLUMN_TITLES = {
-  model: 'AI model summarized in this row. Each row combines that model’s overall value preference and its repeat-pattern mix across repeated conditions in this analysis.',
-  preferredValue: 'The value this model most often favors overall in this analysis slice. This is the top value on the model-level preference summary, not a count of individual transcript decisions.',
-  winRate: 'Win rate for the preferred value. Higher values mean the model chooses that value more often when it is on the selected side.',
-  valueAgreement: 'How often repeated judgments stay on the same value side. Higher means the model usually leans the same way when the same conflict is repeated, even if the exact score changes a little.',
-  stable: 'Share of repeated conditions where the model shows a settled pattern. These are repeats where one clear pattern wins and the answers do not move around much.',
-  softLean: 'Share of repeated conditions where the model shows a narrow but only slightly off-neutral lean. These are coherent repeats with a mild lean, but not strong enough to count as fully settled.',
-  torn: 'Share of repeated conditions where the value conflict remains unresolved between the two sides. These are repeats that stay near the middle or split between sides without a clear winner.',
-  unstable: 'Share of repeated conditions where the answers swing too widely to read as one coherent pattern. These are the broadest or messiest repeats in the set.',
-} as const;
-
-type RepeatPatternMetrics =
-  | {
-      status: 'available';
-      counts: Record<RepeatPattern, number>;
-      conditionIds: Record<RepeatPattern, string[]>;
-      classifiedCount: number;
-      repeatedCount: number;
-      strongerConfidenceCount: number;
-      sourceCount: number;
-    }
-  | {
-      status: 'unavailable';
-      reason: string;
-      repeatedCount: number;
-      strongerConfidenceCount: number;
-      sourceCount: number;
-    };
-
-function buildConditionRows(
-  scenarioDimensions: VisualizationData['scenarioDimensions'] | null | undefined,
-  attributeA: string,
-  attributeB: string,
-): ConditionRow[] {
-  if (!scenarioDimensions || !attributeA || !attributeB) {
-    return [];
-  }
-
-  const grouped = new Map<string, ConditionRow>();
-  Object.entries(scenarioDimensions).forEach(([scenarioId, dimensions]) => {
-    const aLevel = String(dimensions[attributeA] ?? 'N/A');
-    const bLevel = String(dimensions[attributeB] ?? 'N/A');
-    const id = `${aLevel}||${bLevel}`;
-    const current = grouped.get(id);
-    if (current) {
-      current.scenarioIds.push(scenarioId);
-      return;
-    }
-
-    grouped.set(id, {
-      id,
-      attributeALevel: aLevel,
-      attributeBLevel: bLevel,
-      scenarioIds: [scenarioId],
-    });
-  });
-
-  return [...grouped.values()].sort((left, right) => {
-    if (left.attributeALevel === right.attributeALevel) {
-      return left.attributeBLevel.localeCompare(right.attributeBLevel);
-    }
-    return left.attributeALevel.localeCompare(right.attributeALevel);
-  });
-}
-
-function formatPercent(value: number): string {
-  const percentage = value * 100;
-  const roundedToTenth = Math.round(percentage * 10) / 10;
-
-  if (Number.isInteger(roundedToTenth)) {
-    return `${roundedToTenth.toFixed(0)}%`;
-  }
-
-  return `${roundedToTenth.toFixed(1)}%`;
-}
-
-function getPreferredValueName(model: PreferenceViewModel): string | null {
-  return model.topPrioritizedValues[0]?.name
-    ?? model.neutralValues[0]?.name
-    ?? model.topDeprioritizedValues[0]?.name
-    ?? null;
-}
-
-function getPreferenceUnavailableReason(model: PreferenceViewModel): string {
-  return model.availability.status === 'unavailable'
-    ? model.availability.message
-    : 'Preference summary is unavailable for this model.';
-}
-
-function getMetricUnavailableReason(model: ReliabilityViewModel): string {
-  return model.availability.status === 'unavailable'
-    ? model.availability.message
-    : 'This metric is unavailable for this model.';
-}
-
-function InfoTooltipTrigger({
-  label,
-  title,
-}: {
-  label: string;
-  title: string;
-}) {
-  return (
-    <Tooltip
-      content={<div className="max-w-xs whitespace-normal text-xs leading-5">{title}</div>}
-      position="top"
-      variant="light"
-      className="max-w-xs whitespace-normal"
-    >
-      {/* eslint-disable-next-line react/forbid-elements -- Lightweight tooltip trigger requires custom icon-only control */}
-      <button
-        type="button"
-        className="inline-flex cursor-help text-gray-400 transition-colors hover:text-gray-600 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-1 rounded-sm"
-        aria-label={`${label}: ${title}`}
-      >
-        <Info className="h-3.5 w-3.5" />
-      </button>
-    </Tooltip>
-  );
-}
-
-function SummaryHeader({
-  label,
-  title,
-  align = 'left',
-}: {
-  label: string;
-  title: string;
-  align?: 'left' | 'center';
-}) {
-  return (
-    <div className={`flex items-center gap-1 ${align === 'center' ? 'justify-center' : 'justify-start'}`}>
-      <span>{label}</span>
-      <InfoTooltipTrigger label={label} title={title} />
-    </div>
-  );
-}
-
-function getOverviewUnavailableMessage(
-  semantics: AnalysisSemanticsView,
-  analysisMode?: 'single' | 'paired',
-  hasCompanionAnalysis?: boolean,
-): string | null {
-  if (semantics.preference.rowAvailability.status === 'unavailable') {
-    return semantics.preference.rowAvailability.message;
-  }
-
-  if (semantics.reliability.rowAvailability.status !== 'unavailable') {
-    return null;
-  }
-
-  if (
-    analysisMode === 'paired'
-    && hasCompanionAnalysis
-    && semantics.reliability.rowAvailability.reason === 'no-repeat-coverage'
-  ) {
-    return 'This paired view has no repeated measurements per condition, so baseline reliability is unavailable. Pooling both vignette orders does not add repeat coverage.';
-  }
-
-  return semantics.reliability.rowAvailability.message;
-}
-
-function ModeAvailabilitySection({
-  title,
-  message,
-}: {
-  title: string;
-  message: string;
-}) {
-  return (
-    <section className="space-y-1 rounded-lg border border-gray-200 bg-gray-50 p-4">
-      <h4 className="text-sm font-semibold uppercase tracking-wide text-gray-700">{title}</h4>
-      <p className="text-sm text-gray-600">{message}</p>
-    </section>
-  );
-}
-
-function classifyRepeatPattern(
-  directionalAgreement: number | null | undefined,
-  medianSignedDistance: number | null | undefined,
-  neutralShare: number | null | undefined,
-  range: number | null | undefined,
-): RepeatPattern | null {
-  if (
-    directionalAgreement === null
-    || directionalAgreement === undefined
-    || medianSignedDistance === null
-    || medianSignedDistance === undefined
-    || neutralShare === null
-    || neutralShare === undefined
-    || range === null
-    || range === undefined
-  ) {
-    return null;
-  }
-
-  const absoluteDistance = Math.abs(medianSignedDistance);
-  if (directionalAgreement >= 0.80) {
-    return 'stable';
-  }
-  if (absoluteDistance >= 0.50 && range <= 1 && directionalAgreement >= 0.55) {
-    return 'softLean';
-  }
-  if (range >= 3) {
-    return 'noisy';
-  }
-  if (neutralShare >= 0.60 || absoluteDistance < 0.35) {
-    return 'torn';
-  }
-  return 'torn';
-}
-
-function getConditionRepeatStats(
-  modelId: string,
-  scenarioIds: string[],
-  varianceAnalysis: VarianceAnalysis | null | undefined,
-): ConditionRepeatStats | null {
-  if (!varianceAnalysis) return null;
-
-  const modelStats = varianceAnalysis.perModel[modelId];
-  if (!modelStats?.perScenario) return null;
-
-  const repeatedScenarioStats = scenarioIds
-    .map((scenarioId) => modelStats.perScenario[scenarioId])
-    .filter((stats): stats is NonNullable<typeof stats> => stats != null && stats.sampleCount >= 2);
-
-  if (repeatedScenarioStats.length === 0) {
-    return null;
-  }
-
-  const totalCount = repeatedScenarioStats.reduce((sum, stats) => sum + stats.sampleCount, 0);
-  const weightedMean = <T extends keyof (typeof repeatedScenarioStats)[number]>(key: T): number | null => {
-    const populated = repeatedScenarioStats.filter((stats) => typeof stats[key] === 'number');
-    if (populated.length === 0) {
-      return null;
-    }
-
-    const weightedValue = populated.reduce((sum, stats) => {
-      return sum + (Number(stats[key]) * stats.sampleCount);
-    }, 0);
-    const weightedCount = populated.reduce((sum, stats) => sum + stats.sampleCount, 0);
-
-    return weightedCount > 0 ? Number((weightedValue / weightedCount).toFixed(2)) : null;
-  };
-
-  return {
-    directionalAgreement: weightedMean('directionalAgreement'),
-    medianSignedDistance: weightedMean('medianSignedDistance'),
-    neutralShare: weightedMean('neutralShare'),
-    totalCount,
-    maxRange: repeatedScenarioStats.reduce((max, stats) => Math.max(max, stats.range ?? 0), 0),
-  };
-}
-
-function getRepeatPatternMetrics(
-  modelId: string,
-  varianceAnalysis: VarianceAnalysis | null | undefined,
-  conditionRows: ConditionRow[],
-): RepeatPatternMetrics {
-  if (!varianceAnalysis?.perModel[modelId]?.perScenario) {
-    return {
-      status: 'unavailable',
-      reason: 'No repeat data is available for this model.',
-      repeatedCount: 0,
-      strongerConfidenceCount: 0,
-      sourceCount: 1,
-    };
-  }
-
-  if (conditionRows.length === 0) {
-    return {
-      status: 'unavailable',
-      reason: 'Condition-level grouping is unavailable for this run.',
-      repeatedCount: 0,
-      strongerConfidenceCount: 0,
-      sourceCount: 1,
-    };
-  }
-
-  const counts: Record<RepeatPattern, number> = {
-    stable: 0,
-    softLean: 0,
-    torn: 0,
-    noisy: 0,
-  };
-  const conditionIds: Record<RepeatPattern, string[]> = {
-    stable: [],
-    softLean: [],
-    torn: [],
-    noisy: [],
-  };
-  let strongerConfidenceCount = 0;
-  let repeatedCount = 0;
-
-  conditionRows.forEach((row) => {
-    const conditionStats = getConditionRepeatStats(modelId, row.scenarioIds, varianceAnalysis);
-    if (!conditionStats) {
-      return;
-    }
-
-    repeatedCount += 1;
-    if (conditionStats.totalCount >= 10) {
-      strongerConfidenceCount += 1;
-    }
-
-    const pattern = classifyRepeatPattern(
-      conditionStats.directionalAgreement,
-      conditionStats.medianSignedDistance,
-      conditionStats.neutralShare,
-      conditionStats.maxRange,
-    );
-    if (!pattern) {
-      return;
-    }
-
-    counts[pattern] += 1;
-    conditionIds[pattern].push(row.id);
-  });
-
-  const classifiedCount = counts.stable + counts.softLean + counts.torn + counts.noisy;
-  if (classifiedCount === 0) {
-    return {
-      status: 'unavailable',
-      reason: 'Some repeat data is present, but not enough conditions produced a publishable stability classification.',
-      repeatedCount,
-      strongerConfidenceCount,
-      sourceCount: 1,
-    };
-  }
-
-  return {
-    status: 'available',
-    counts,
-    conditionIds,
-    classifiedCount,
-    repeatedCount,
-    strongerConfidenceCount,
-    sourceCount: 1,
-  };
-}
-
-function mergeRepeatPatternMetrics(
-  metricsList: RepeatPatternMetrics[],
-): RepeatPatternMetrics {
-  if (metricsList.length === 0) {
-    return {
-      status: 'unavailable',
-      reason: 'No repeat data is available for this view.',
-      repeatedCount: 0,
-      strongerConfidenceCount: 0,
-      sourceCount: 0,
-    };
-  }
-
-  const availableMetrics = metricsList.filter(
-    (metrics): metrics is Extract<RepeatPatternMetrics, { status: 'available' }> => metrics.status === 'available',
-  );
-  const repeatedCount = metricsList.reduce((sum, metrics) => sum + metrics.repeatedCount, 0);
-  const strongerConfidenceCount = metricsList.reduce((sum, metrics) => sum + metrics.strongerConfidenceCount, 0);
-  const sourceCount = metricsList.reduce((sum, metrics) => sum + metrics.sourceCount, 0);
-
-  if (availableMetrics.length === 0) {
-    const firstReason = metricsList.find(
-      (metrics): metrics is Extract<RepeatPatternMetrics, { status: 'unavailable' }> => metrics.status === 'unavailable',
-    )?.reason;
-    return {
-      status: 'unavailable',
-      reason: firstReason ?? 'No repeat data is available for this view.',
-      repeatedCount,
-      strongerConfidenceCount,
-      sourceCount,
-    };
-  }
-
-  const counts: Record<RepeatPattern, number> = {
-    stable: 0,
-    softLean: 0,
-    torn: 0,
-    noisy: 0,
-  };
-  const conditionIds: Record<RepeatPattern, string[]> = {
-    stable: [],
-    softLean: [],
-    torn: [],
-    noisy: [],
-  };
-
-  availableMetrics.forEach((metrics) => {
-    (['stable', 'softLean', 'torn', 'noisy'] as const).forEach((pattern) => {
-      counts[pattern] += metrics.counts[pattern];
-      conditionIds[pattern].push(...metrics.conditionIds[pattern]);
-    });
-  });
-
-  return {
-    status: 'available',
-    counts,
-    conditionIds,
-    classifiedCount: availableMetrics.reduce((sum, metrics) => sum + metrics.classifiedCount, 0),
-    repeatedCount,
-    strongerConfidenceCount,
-    sourceCount,
-  };
-}
-
-function SummaryCell({
-  children,
-  title,
-  showInfoIcon = false,
-  align = 'left',
-}: {
-  children: ReactNode;
-  title?: string;
-  showInfoIcon?: boolean;
-  align?: 'left' | 'center';
-}) {
-  return (
-    <div className={`flex items-center gap-1 truncate ${align === 'center' ? 'justify-center' : ''}`}>
-      <div className="truncate" aria-label={title}>
-        {children}
-      </div>
-      {showInfoIcon && title ? <InfoTooltipTrigger label="Cell details" title={title} /> : null}
-    </div>
-  );
-}
-
-function PatternMetricButton({
-  runId,
-  analysisBasePath,
-  analysisSearchParams,
-  modelId,
-  pattern,
-  metrics,
-  title,
-  rowDim,
-  colDim,
-}: {
-  runId: string;
-  analysisBasePath: AnalysisBasePath;
-  analysisSearchParams?: URLSearchParams | string;
-  modelId: string;
-  pattern: RepeatPattern;
-  metrics: Extract<RepeatPatternMetrics, { status: 'available' }>;
-  title: string;
-  rowDim: string;
-  colDim: string;
-}) {
-  const navigate = useNavigate();
-  const count = metrics.counts[pattern];
-  const value = metrics.classifiedCount === 0 ? 0 : count / metrics.classifiedCount;
-
-  if (count === 0) {
-    return <SummaryCell title={title} align="center">{formatPercent(value)}</SummaryCell>;
-  }
-
-  return (
-    <Button
-      type="button"
-      variant="ghost"
-      size="sm"
-      className="h-auto min-h-0 px-0 py-0 text-sm font-medium text-gray-700 hover:bg-transparent hover:text-teal-700"
-      title={title}
-      aria-label={title}
-      onClick={() => {
-        const params = new URLSearchParams({
-          modelId,
-          repeatPattern: pattern,
-          rowDim,
-          colDim,
-          conditionIds: metrics.conditionIds[pattern].join(','),
-        });
-        navigate(buildAnalysisTranscriptsPath(analysisBasePath, runId, params, analysisSearchParams));
-      }}
-    >
-      {formatPercent(value)}
-    </Button>
-  );
-}
-
-function PairedPatternMetricButton({
-  runId,
-  companionRunId,
-  analysisBasePath,
-  analysisSearchParams,
-  modelId,
-  pattern,
-  primaryMetrics,
-  companionMetrics,
-  title,
-  rowDim,
-  colDim,
-}: {
-  runId: string;
-  companionRunId: string;
-  analysisBasePath: AnalysisBasePath;
-  analysisSearchParams?: URLSearchParams | string;
-  modelId: string;
-  pattern: RepeatPattern;
-  primaryMetrics: Extract<RepeatPatternMetrics, { status: 'available' }>;
-  companionMetrics: Extract<RepeatPatternMetrics, { status: 'available' }>;
-  title: string;
-  rowDim: string;
-  colDim: string;
-}) {
-  const navigate = useNavigate();
-  const primaryCount = primaryMetrics.counts[pattern];
-  const companionCount = companionMetrics.counts[pattern];
-  const totalClassified = primaryMetrics.classifiedCount + companionMetrics.classifiedCount;
-  const totalCount = primaryCount + companionCount;
-  const value = totalClassified === 0 ? 0 : totalCount / totalClassified;
-
-  if (totalCount === 0) {
-    return <SummaryCell title={title} align="center">{formatPercent(value)}</SummaryCell>;
-  }
-
-  return (
-    <Button
-      type="button"
-      variant="ghost"
-      size="sm"
-      className="h-auto min-h-0 px-0 py-0 text-sm font-medium text-gray-700 hover:bg-transparent hover:text-teal-700"
-      title={title}
-      aria-label={title}
-      onClick={() => {
-        const params = new URLSearchParams({
-          modelId,
-          repeatPattern: pattern,
-          rowDim,
-          colDim,
-          companionRunId,
-          primaryConditionIds: primaryMetrics.conditionIds[pattern].join(','),
-          companionConditionIds: companionMetrics.conditionIds[pattern].join(','),
-        });
-        navigate(buildAnalysisTranscriptsPath(analysisBasePath, runId, params, analysisSearchParams));
-      }}
-    >
-      {formatPercent(value)}
-    </Button>
-  );
-}
-
-function OverviewSummaryTable({
-  runId,
-  analysisBasePath = ANALYSIS_BASE_PATH,
-  analysisSearchParams,
-  semantics,
-  varianceAnalysis,
-  visualizationData,
-  companionAnalysis,
-  expectedAttributes = [],
-  completedBatches,
-  aggregateSourceRunCount,
-  coverageBatchCount,
-  coveragePairedBatchCount,
-  isAggregate,
-  analysisMode,
-  currentRun,
-  currentAnalysis,
-  companionRun,
-}: {
-  runId: string;
-  analysisBasePath?: AnalysisBasePath;
-  analysisSearchParams?: URLSearchParams | string;
-  semantics: AnalysisSemanticsView;
-  varianceAnalysis?: VarianceAnalysis | null;
-  visualizationData: VisualizationData | null | undefined;
-  companionAnalysis?: AnalysisResult | null;
-  expectedAttributes?: string[];
-  completedBatches: number | '-';
-  aggregateSourceRunCount: number | null;
-  coverageBatchCount?: number | null;
-  coveragePairedBatchCount?: number | null;
-  isAggregate: boolean;
-  analysisMode?: 'single' | 'paired';
-  currentRun?: Run | null;
-  currentAnalysis?: AnalysisResult | null;
-  companionRun?: Run | null;
-}) {
-  const models = useMemo(() => {
-    return Object.keys(semantics.preference.byModel)
-      .sort((left, right) => left.localeCompare(right))
-      .map((modelId) => {
-        const preference = semantics.preference.byModel[modelId];
-        const reliability = semantics.reliability.byModel[modelId];
-        if (!preference || !reliability) {
-          return null;
-        }
-
-        return {
-          modelId,
-          preference,
-          reliability,
-        };
-      })
-      .filter((model): model is {
-        modelId: string;
-        preference: PreferenceViewModel;
-        reliability: ReliabilityViewModel;
-      } => model !== null);
-  }, [semantics.preference.byModel, semantics.reliability.byModel]);
-  const scenarioDimensions = visualizationData?.scenarioDimensions;
-  const modelScenarioMatrix = visualizationData?.modelScenarioMatrix;
-  const availableAttributes = useMemo(() => {
-    return resolveScenarioAttributes(scenarioDimensions, expectedAttributes, modelScenarioMatrix);
-  }, [expectedAttributes, modelScenarioMatrix, scenarioDimensions]);
-  const attributeA = availableAttributes[0] ?? '';
-  const attributeB = availableAttributes[1] ?? availableAttributes[0] ?? '';
-  const conditionRows = useMemo(
-    () => buildConditionRows(scenarioDimensions, attributeA, attributeB),
-    [attributeA, attributeB, scenarioDimensions],
-  );
-  const companionConditionRows = useMemo(
-    () => buildConditionRows(companionAnalysis?.visualizationData?.scenarioDimensions, attributeA, attributeB),
-    [attributeA, attributeB, companionAnalysis?.visualizationData?.scenarioDimensions],
-  );
-  const repeatPatternSources = useMemo<RepeatPatternSource[]>(() => {
-    const sources: RepeatPatternSource[] = [{
-      runId,
-      varianceAnalysis,
-      conditionRows,
-    }];
-
-    if (analysisMode === 'paired' && companionAnalysis) {
-      sources.push({
-        runId: companionAnalysis.runId,
-        varianceAnalysis: companionAnalysis.varianceAnalysis,
-        conditionRows: companionConditionRows,
-      });
-    }
-
-    return sources;
-  }, [analysisMode, companionAnalysis, companionConditionRows, conditionRows, runId, varianceAnalysis]);
-  const isPooledAcrossRuns = repeatPatternSources.length > 1;
-  const summaryUnavailableMessage = getOverviewUnavailableMessage(
-    semantics,
-    analysisMode,
-    companionAnalysis != null,
-  );
-
-  const helperText = coverageBatchCount !== null && coverageBatchCount !== undefined
-    ? coveragePairedBatchCount !== null && coveragePairedBatchCount !== undefined
-      ? `Run-level evidence: ${coverageBatchCount} batches from coverage cell • ${coveragePairedBatchCount} paired batches`
-      : `Run-level evidence: ${coverageBatchCount} batches from coverage cell`
-    : analysisMode === 'paired' && companionAnalysis
-      ? `Run-level evidence: pooled across ${repeatPatternSources.length} companion runs`
-      : isAggregate
-        ? aggregateSourceRunCount === null
-          ? 'Run-level evidence: contributing source-run count unavailable'
-          : `Run-level evidence: ${aggregateSourceRunCount} contributing source run${aggregateSourceRunCount === 1 ? '' : 's'}`
-        : completedBatches === '-'
-          ? 'Run-level evidence: completed batch count unavailable'
-          : `Run-level evidence: ${completedBatches} completed batch${completedBatches === 1 ? '' : 'es'}`;
-  const summaryRef = useRef<HTMLDivElement>(null);
-  return (
-    <div ref={summaryRef} className="space-y-3 rounded-lg border border-gray-200 p-4">
-      <div className="flex flex-wrap items-start justify-between gap-3">
-        <div>
-          <h3 className="text-sm font-medium text-gray-700">Overview Summary</h3>
-          <p className="mt-1 text-xs text-gray-500">{helperText}</p>
-        </div>
-        <CopyVisualButton targetRef={summaryRef} label="overview summary" />
-      </div>
-
-      {summaryUnavailableMessage ? (
-        <div className="rounded-md border border-amber-200 bg-amber-50 p-3 text-sm text-amber-700">
-          {summaryUnavailableMessage}
-        </div>
-      ) : null}
-
-      <div className="overflow-x-auto">
-        <table className="min-w-full border-collapse">
-          <thead>
-            <tr>
-              <th className="border border-gray-200 bg-gray-50 px-3 py-2 text-left text-xs font-semibold uppercase text-gray-600">
-                <SummaryHeader label="Model" title={SUMMARY_COLUMN_TITLES.model} />
-              </th>
-              <th className="border border-gray-200 bg-gray-50 px-3 py-2 text-left text-xs font-semibold uppercase text-gray-600">
-                <SummaryHeader label="Preferred Value" title={SUMMARY_COLUMN_TITLES.preferredValue} />
-              </th>
-              <th className="border border-gray-200 bg-gray-50 px-3 py-2 text-left text-xs font-semibold uppercase text-gray-600">
-                <SummaryHeader label="Win Rate" title={SUMMARY_COLUMN_TITLES.winRate} />
-              </th>
-              <th className="border border-gray-200 bg-gray-50 px-3 py-2 text-center text-xs font-semibold uppercase text-gray-600">
-                <SummaryHeader label="Value Agreement" title={SUMMARY_COLUMN_TITLES.valueAgreement} align="center" />
-              </th>
-              <th className="border border-gray-200 bg-gray-50 px-3 py-2 text-center text-xs font-semibold uppercase text-gray-600">
-                <SummaryHeader label="Stable %" title={SUMMARY_COLUMN_TITLES.stable} align="center" />
-              </th>
-              <th className="border border-gray-200 bg-gray-50 px-3 py-2 text-center text-xs font-semibold uppercase text-gray-600">
-                <SummaryHeader label="Soft Lean %" title={SUMMARY_COLUMN_TITLES.softLean} align="center" />
-              </th>
-              <th className="border border-gray-200 bg-gray-50 px-3 py-2 text-center text-xs font-semibold uppercase text-gray-600">
-                <SummaryHeader label="Torn %" title={SUMMARY_COLUMN_TITLES.torn} align="center" />
-              </th>
-              <th className="border border-gray-200 bg-gray-50 px-3 py-2 text-center text-xs font-semibold uppercase text-gray-600">
-                <SummaryHeader label="Unstable %" title={SUMMARY_COLUMN_TITLES.unstable} align="center" />
-              </th>
-            </tr>
-          </thead>
-          <tbody>
-            {models.map(({ modelId, preference, reliability }) => {
-              const preferredValue = getPreferredValueName(preference);
-              const preferredValueWinRate = preference.topPrioritizedValues[0]?.winRate ?? null;
-              const perSourceMetricsList = repeatPatternSources.map((source) => getRepeatPatternMetrics(modelId, source.varianceAnalysis, source.conditionRows));
-              const repeatMetrics = mergeRepeatPatternMetrics(perSourceMetricsList);
-              const primarySourceMetrics = perSourceMetricsList[0] ?? null;
-              const companionSourceMetrics = perSourceMetricsList[1] ?? null;
-              const repeatEvidenceDetail = repeatMetrics.status === 'available'
-                ? `${repeatMetrics.classifiedCount} of ${repeatMetrics.repeatedCount} repeated conditions classified • ${repeatMetrics.strongerConfidenceCount} condition${repeatMetrics.strongerConfidenceCount === 1 ? '' : 's'} with 10+ repeats`
-                : `${repeatMetrics.reason} • ${repeatMetrics.strongerConfidenceCount} condition${repeatMetrics.strongerConfidenceCount === 1 ? '' : 's'} with 10+ repeats`;
-
-              return (
-                <tr key={modelId}>
-                  <td className="border border-gray-200 px-3 py-2 text-sm text-gray-700">
-                    <div className="font-medium text-gray-900">{modelId}</div>
-                  </td>
-                  <td className="border border-gray-200 px-3 py-2 text-sm text-gray-700">
-                    {preferredValue ? (
-                      <SummaryCell title={preferredValue}>{preferredValue}</SummaryCell>
-                    ) : (
-                      <SummaryCell title={getPreferenceUnavailableReason(preference)} showInfoIcon>—</SummaryCell>
-                    )}
-                  </td>
-                  <td className="border border-gray-200 px-3 py-2 text-sm text-gray-700">
-                    {preferredValueWinRate !== null && preferredValueWinRate !== undefined ? (
-                      <SummaryCell title={`Preferred value win rate: ${Math.round(preferredValueWinRate * 100)}%`}>
-                        {Math.round(preferredValueWinRate * 100)}%
-                      </SummaryCell>
-                    ) : (
-                      <SummaryCell title={getPreferenceUnavailableReason(preference)} showInfoIcon>—</SummaryCell>
-                    )}
-                  </td>
-                  <td className="border border-gray-200 px-3 py-2 text-center text-sm text-gray-700">
-                    {reliability.directionalAgreement !== null ? (
-                      <SummaryCell title={`Directional agreement: ${formatPercent(reliability.directionalAgreement)}`}>
-                        {formatPercent(reliability.directionalAgreement)}
-                      </SummaryCell>
-                    ) : (
-                      <SummaryCell title={getMetricUnavailableReason(reliability)} showInfoIcon>—</SummaryCell>
-                    )}
-                  </td>
-                  {(['stable', 'softLean', 'torn', 'noisy'] as const).map((pattern) => {
-                    const label = REPEAT_PATTERN_LABELS[pattern];
-                    const count = repeatMetrics.status === 'available' ? repeatMetrics.counts[pattern] : 0;
-                    const title = repeatMetrics.status === 'available'
-                      ? isPooledAcrossRuns
-                        ? `${label}: ${repeatMetrics.counts[pattern]} of ${repeatMetrics.classifiedCount} repeated conditions across both vignette orders`
-                        : `${label}: ${repeatMetrics.counts[pattern]} of ${repeatMetrics.classifiedCount} repeated conditions • ${repeatMetrics.strongerConfidenceCount} condition${repeatMetrics.strongerConfidenceCount === 1 ? '' : 's'} with 10+ repeats`
-                      : repeatEvidenceDetail;
-
-                    return (
-                      <td key={pattern} className="border border-gray-200 px-3 py-2 text-center text-sm text-gray-700">
-                        {repeatMetrics.status === 'available' && count > 0 && !isPooledAcrossRuns ? (
-                          <PatternMetricButton
-                            runId={runId}
-                            analysisBasePath={analysisBasePath}
-                            analysisSearchParams={analysisSearchParams}
-                            modelId={modelId}
-                            pattern={pattern}
-                            metrics={repeatMetrics}
-                            title={title}
-                            rowDim={attributeA}
-                            colDim={attributeB}
-                          />
-                        ) : repeatMetrics.status === 'available' && count > 0 && isPooledAcrossRuns
-                            && companionAnalysis != null
-                            && primarySourceMetrics !== null && primarySourceMetrics.status === 'available'
-                            && companionSourceMetrics !== null && companionSourceMetrics.status === 'available' ? (
-                          <PairedPatternMetricButton
-                            runId={runId}
-                            companionRunId={companionAnalysis.runId}
-                            analysisBasePath={analysisBasePath}
-                            analysisSearchParams={analysisSearchParams}
-                            modelId={modelId}
-                            pattern={pattern}
-                            primaryMetrics={primarySourceMetrics}
-                            companionMetrics={companionSourceMetrics}
-                            title={title}
-                            rowDim={attributeA}
-                            colDim={attributeB}
-                          />
-                        ) : (
-                          <SummaryCell title={title} showInfoIcon={repeatMetrics.status === 'unavailable'} align="center">
-                            {repeatMetrics.status === 'available'
-                              ? formatPercent(repeatMetrics.classifiedCount === 0 ? 0 : repeatMetrics.counts[pattern] / repeatMetrics.classifiedCount)
-                              : '—'}
-                          </SummaryCell>
-                        )}
-                      </td>
-                    );
-                  })}
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      </div>
-
-      <div className="border-t border-gray-200 pt-4">
-        {analysisMode === 'paired' && currentRun && currentAnalysis ? (
-          <PairedRunComparisonCard
-            currentRun={currentRun}
-            currentAnalysis={currentAnalysis}
-            companionRun={companionRun ?? null}
-            companionAnalysis={companionAnalysis ?? null}
-            analysisBasePath={analysisBasePath}
-            analysisSearch={typeof analysisSearchParams === 'string' ? analysisSearchParams : analysisSearchParams?.toString() ?? ''}
-            embedded
-          />
-        ) : (
-          <ModeAvailabilitySection
-            title="Paired Run Comparison"
-            message="This comparison is only available in Paired vignettes mode. Switch the toggle above to see both vignette orders together."
-          />
-        )}
-      </div>
-    </div>
-  );
-}
 
 export function OverviewTab({
   runId,
